move test to inmem bucket
